Implement skipping of Windows symbols which are string or number constants.
